pyproject.toml
[project]
name = "airos-agent-demo"
version = "0.1.0"
description = "Demo agent showcasing all AirOS SDK features"
requires-python = ">=3.10"
dependencies = [
    "airos-sdk>=0.4.0",
    "groq>=0.4.0",
]
